requirements.txt
openai>=1.65,<2
discord.py>=2.4,<3
python-dotenv>=1.0.1,<2
httpx>=0.27,<1
beautifulsoup4>=4.12,<5
feedparser>=6.0,<7
